requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
axios==0.4.0
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-ses==4.1.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
jiter==0.6.1
jmespath==1.0.1
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
openai==1.52.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
